Remove YACReader3DImage::index data member
This index was used only in YACReaderFlowGL::drawCover(). A simple
change of this function eliminates the use.

The benefits of the removal:
* sizeof(YACReader3DImage) is 8 bytes less now, which improves CPU cache
  utilization and saves a little RAM;
* the code, which ensured that this index member matches the index of
  the image in YACReaderFlowGL::images, can be removed now;
* the index member can no longer get out of sync with the index of the
  image.

Simplify the implementation of YACReaderFlowGL::draw().

For some (most likely obsolete) reason YACReaderFlowGL::replace()
checked if the index member matches the index of the image and handled
the mismatch differently. I don't see how the mismatch could happen with
the current implementation of YACReaderFlowGL and I couldn't trigger it
during debugging. So I think this special case can be safely removed.