YACReaderFlow: use ScrollManagement to handle wheelEvent()
GoToFlow and ComicFlow are the only two classes that use YACReaderFlow.
They both handled wheelEvent() manually in practically the same way as
YACReaderFlow does now, except for time and minimum move checks in
ScrollManagement::getMovement().

This new implementation is equivalent to and consistent with the OpenGL
version in YACReaderFlowGL::wheelEvent().