Port ScrollManagement and Viewer away from deprecated QWheelEvent API
QWheelEvent::orientation() and QWheelEvent::delta() are deprecated. The
documentation for both suggests using QWheelEvent::angleDelta() instead.

Keep treating the horizontal mouse wheel the same as vertical in Comic
and Page Flow: comics and pages move horizontally, so horizontal
scrolling is even more natural than vertical in these cases.

Viewer handles horizontal scrolling separately - to scroll the comic
page to the right or to the left. So ScrollManagement must ignore
horizontal delta in this case. For this purpose the new parameter in
ScrollManagement::getMovement() determines which orientations are taken
into account: horizontal, vertical or both.

Remove redundant parentheses in long `if` statements to improve
readability.