Library: fix behavior after assigning current order to comics
After the current order was assigned to comics while a reading list was
selected or in search mode, the last selected folder's info was loaded.
In the search mode case this made the UI look half-way between Normal
and Searching navigation statuses. Now the last selected folder's or
list's info or the last search result is reloaded.

One issue remains: after the current order is assigned in search mode,
the first comic in the search results is selected rather than the comic
to which the first number was assigned. I don't know how to fix this.

An alternative fix is to disable asignOrderAction while a reading list
is selected and in search mode. But this is more difficult to implement
and inconsistent with the other enabled-in-these-cases comic actions.
Disabling the action would also remove the existing possibly useful
feature of assigning numbers across directories.