LibraryWindow::deleteComics(): don't rely on selection
This commit fixes two bugs:
1. When the current reading list is deselected by clicking outside of
all rows in YACReaderReadingListsView, the reading list is still loaded
and active in the comics view. Triggering the *Delete selected comics*
action deletes the selected comics from disk rather than from the
current label/list then.
2. The meaning of the *Delete selected comics* action in search mode
depends on whether a reading list was (and still is) selected when the
search mode was entered.

Now the comics are deleted from the current label/list if and only if
a reading list is loaded in the comics view. It doesn't make sense to
delete comics from some reading list in search mode, so deletion from
disk occurs, as well as when a folder is loaded in the comics view.