Don't reselect current source needlessly
Wherever YACReaderNavigationController::reselectCurrentSource() was
called, simply reloading the comics view is sufficient: when a dialog is
accepted or when the last comic is deleted from the current reading
list. Similarly there is no need to reselect the current folder when the
last comic is deleted from the current folder.

Simply reloading the comics view saves some work. In addition, in all
changed places forward navigation history entries were cut off and lost
because of the call to YACReaderHistoryController::updateHistory() in
selectedFolder() or selectedList().

Now that reselectCurrentSource() is not used anymore, it can be safely
removed rather than fixed. Its implementation suffered from two bugs:
1. Checking whether a reading list is selected is not reliable, because
a reading list can be deselected but not unloaded by clicking outside of
all rows in YACReaderReadingListsView.
2. It didn't account for the possibility of active search mode.