Don't crash during a slow (>= 60 seconds) startup on X11
The corresponding upstream bug - QTBUG-58709 - was reported in 2017 and
was de-prioritized from P2 to P3 in 2020.

A quick benchmark shows that the added QCoreApplication::processEvents()
call takes the same time in Debug and Release builds of YACReader and
YACReaderLibrary - from 0.5 to 0.6 milliseconds. 0.6 milliseconds is not
a noticeable startup slowdown, especially considering that the event
loop does useful work: processes events, which would have to be handled
eventually anyway.

Don't restrict the workaround to Linux/X11 or Debug mode: no need to
introduce a potential behavior difference between supported platforms
or build types.

The same workaround can be applied to YACReaderLibraryServer, but I
don't use the server and cannot test it.